+++ lib/pluck_map/attribute_builder.rb
@@ -1,5 +1,6 @@
 require "pluck_map/attribute"
 require "pluck_map/attributes"
+require "pluck_map/relationships"
 
 module PluckMap
   class AttributeBuilder < BasicObject
@@ -12,7 +13,7 @@ module PluckMap
       else
         builder.instance_eval(&block)
       end
-      Attributes.new(attributes)
+      Attributes.new(attributes, model)
     end
 
     def initialize(attributes, model)
@@ -25,6 +26,20 @@ module PluckMap
       options[:value] = args.first unless args.empty?
       @attributes.push Attribute.new(attribute_name, @model, options)
       :attribute_added
+    end
+
+    def has_many(name, *args, &block)
+      options = args.extract_options!
+      options[:scope_block] = args.first unless args.empty?
+      @attributes.push Relationships.many(@model, name, block, options)
+      :relationship_added
+    end
+
+    def has_one(name, *args, &block)
+      options = args.extract_options!
+      options[:scope_block] = args.first unless args.empty?
+      @attributes.push Relationships.one(@model, name, block, options)
+      :relationship_added
     end
 
   end