+++ test/support/character.rb
@@ -1,0 +1,3 @@
+class Character < ActiveRecord::Base
+  has_and_belongs_to_many :books, inverse_of: :characters
+end