+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * BREAKING: Remove deprecated features/methods (@boblail)
 * FEATURE: Optimize `to_json` when a presenter doesn't need to process values in Ruby (@boblail)
+* FEATURE: Add `has_many` and `has_one` DSL for presenting nested resources (@boblail)
 
 ## v0.6.1 (2019 May 12)